refactor status route instance ops
